+++ src/main/java/org/project/spring_mini_project/features/student/StudentRepository.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 
 @Repository
 
-public interface StudentRepository extends JpaRepository<Student, Long>
+public interface StudentRepository extends JpaRepository<Student, Integer>
 {
     Optional<Student> findByUserUsername(String username);
 }
